Make method `TextModel.model_description` concrete
It kreates a skeleton keras.models.Sequential model.
That model has  a single, Dense output layer and relies on the new,
abstract `hidden_layers` method to fill-in the rest of the layers.

This approach makes the models DRY-er.

Conv2DModel does not use a Sequential model, so it overrides the
`model_description` method alltogether and only has a blank, unused
`hidden_layers` method.