Create SequenceModel abstract class
This will be the ancestor for a set of models that consume
sequence-of-words (at the moment SequenceModel inherits
from WordLevelModel so it can not represent char-level models).

SequenceModel handles vectorizing texts as padded sequences and
creating the hidden layers using an `embedding layer` and a list
of `recurrent_layers`.

`recurrent_layers` is abstract (and does not actually have to be
recurrent) and `embedding_layer` provides a basic implementation
but can be overriden (eg. for pre-trained embeddings)