Make vocabulary size be an instance method instead of a constant
Up to now, each WordLevelModel had a class-level constan, VOCAB_SIZE
to signify its vocabulary size.
However, for models trained on smaller corpora, this is wasteful as
a reasonable default for this constant was actually greater as the
corpus's actual vocabulary size.

Terefore the instace method vocabulary_size() is created and
it will return the actual vocabulary size or the maximum vocuabulary
size (MAX_VOCAB_SIZE) if we need to limit it.