Do not use an Embedding layer in CharacterLevelModel
In a naive implementation, at least for text classification,
using an Embedding layer for characters just adds complexity
to the models architecture.
In this use case the vocabulary size is bounded (by the number
of available characters) and (intuitively) the correlations among
tokens are weak to non-existent.

Also empirically removing the Embedding layer yields better
precision.